pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "diveroast"
version = "0.1.0"
description = "Conversational diver roasting agent powered by DAN guidelines and Gemini"
requires-python = ">=3.11"
dependencies = [
    # Core
    "pandas>=2.2",
    "numpy>=1.26",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    # RAG
    "dlt[lancedb]>=0.5",
    "lancedb>=0.13",
    "beautifulsoup4>=4.12",
    "langchain-text-splitters>=0.2",
    # LLM
    "google-genai>=1.0",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0",
    # MCP
    "mcp[cli]>=1.0",
    # Observability
    "arize-phoenix-otel>=0.6",
    "openinference-instrumentation-google-genai>=0.1",
    # Embeddings
    "sentence-transformers>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "pyrefly>=0.1",
    "pre-commit>=3.7",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88
exclude = ["src/rag/rest_api"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
